Solution in Erlang.
Uses dynamic programming as proposed. However we do not use a full supervisor structure.
Instead, I start a separate process that I use as a "solution server", which is basically
a side effect for caching / memoizing solutions.

Currently, I only cache "tasks", not simple solutions like single piece moves, as they
are considered to be trivial anyway.

One optimisation that could be done here is to generalize the solutions so that
for example "move 3 from a to b via c" recognizes the solution "move 3 from b to c via a"
as a viable solution (and possibly rewrite it "live" for the asker)