Okay, could not leave it be.. Now it recognizes similar solutions as well
This is done in the solution server by trying all permutations of the
task at hand, and telling the client whether it was an exact match
or a similar match.

The client also gained some functionality to indent the output, and tell
the user what to do.

Note that we don't store the actual steps anymore, since we're not going
to output them anyway. Let stdout be the memory and let the user look
it up... much faster (for us, certainly not for the user!)